Remove wizard interface
