+++ maven-plugin/src/main/java/org/openmrs/maven/plugins/AbstractTask.java
@@ -76,12 +76,6 @@ public abstract class AbstractTask extends AbstractMojo {
     BuildPluginManager pluginManager;
 
     /**
-     * @component
-     * @required
-     */
-    Wizard wizard;
-
-    /**
      * test mode, if true disables interactive mode and uses batchAnswers, even if there is none
      *
      * @parameter  property="testMode" default-value="false"
@@ -107,6 +101,11 @@ public abstract class AbstractTask extends AbstractMojo {
      * @parameter  property="stats" default-value="false"
      */
     boolean stats;
+
+    /**
+     * the wonderful wiz, if ever a wiz there was
+     */
+    Wizard wizard;
 
     /**
      * wizard for resolving artifact available versions
@@ -164,6 +163,9 @@ public abstract class AbstractTask extends AbstractMojo {
     }
 
     public void initTask() {
+        if (wizard == null) {
+            wizard = new Wizard();
+        }
         if(jira == null){
             jira = new DefaultJira();
         }
